bison 2.4.3 (OS X 10.5 comes with 2.3)
Bison is a general-purpose parser generator that converts an annotated
context-free grammar into an LALR(1) or GLR parser for that grammar. Once
you are proficient with Bison, you can use it to develop a wide range of
language parsers, from those used in simple desk calculators to complex
programming languages.