Add support for gcc-4.5.0 release.
So far I have only tested it on my laptop (Core 2 (x86) MacBook Pro)
compiling for c and c++ only.  It should work for everything but Ada
as I have not looked at the dependencies for that as of yet.  I have
never done anything with Ruby before, so I am sure there are some things
that are going to need to be updated, especially the way I take in options.
On that note I have an option that will build gcc and friends with the -4.5
suffix so it will not mess with other things, but it is not enabled by default.
I checked with otool, and it is taking it, but it is doing so from the Cellar,
and not from /usr/local (which is my prefix).  I did it this way because I
figured that gcc was looking for, and compiled with specific versions of gmp,
mpfr, and mpc, and that it should be linked to that location.  Currently I have
it set up to only compile c and c++, as that is all that I have tested, and did
not want to send out an untested product (gcc defaults to installing everything
but Ada and Obj-C++).  I thought it appropreate to add aliases for both g++ and
gfortran, I also made it so that when the install was envoked from that alias
that it would default to adding fortran to the default list of installed languages.
At the same time, I do not remove c or c++ from the default, and I do this only
because I have not fully looked into what depends on what yet.  Also, I just
compiled llvm/clang with this, and it is running just fine, just as one random
test case

TODO:
Figure out a better way to deal with options
Add support for gcc trunk
Add minimum version number to dependencies (couldn't find how to do that)
Add the ability to run gccs' test suite
Generally test it more to make sure everything is all still working