libsigsegv: version 2.8 and other updates
1. use http rather than ftp.

HTTP is usable for all users, even for those who have only a HTTP_PROXY.

2. update homepage.

The sourceforge one seems out of date and not updated.

3. Build a universal binary by default.

The formula `clisp` requires a 32-bit libsigsegv, but it's not wise
to make libsigsegv `ENV.m32`, so I think build a universal binary
(which is only 8Kb bigger than the 64-bit only one) is a reasonable
choice.